Fix tailing alignment matches
Because an alignment in the middle that opens a gap and completes with
matches at the end is scored just as high as an alignment which
performs all matches before, so an alignment like this appears:

——TCAGT——A
ACTTCAGTAACA

Instead of having the A left aligned all the way.  This change makes it
so given equal scores the gap is selected first.